admin features test reorganization and improvements
