+++ app/policies/stash_engine/proposed_change_policy.rb
@@ -2,11 +2,11 @@ module StashEngine
   class ProposedChangePolicy < ApplicationPolicy
 
     def index?
-      @user.limited_curator?
+      @user.curator?
     end
 
     def update?
-      @user.limited_curator?
+      @user.curator?
     end
 
     def edit?
@@ -14,7 +14,7 @@ module StashEngine
     end
 
     def destroy?
-      @user.limited_curator?
+      @user.curator?
     end
 
   end
